pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-trade"
version = "0.1.0"
description = "Forex Trainer & Paper Trading Dashboard monorepo scaffold"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14,<2.0",
  "celery>=5.4,<6.0",
  "fastapi>=0.115,<1.0",
  "pydantic-settings>=2.6,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "redis>=5.2,<6.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn>=0.32,<1.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "pytest>=8.3,<9.0",
  "ruff>=0.8,<1.0"
]

[project.scripts]
rl-trade-api = "rl_trade_api.main:run"
rl-trade-worker = "rl_trade_worker.main:run"
rl-trade-scheduler = "rl_trade_worker.scheduler:run"

[tool.setuptools.packages.find]
where = [
  "apps/api/src",
  "apps/worker/src",
  "libs/common/src",
  "libs/data/src",
  "libs/features/src",
  "libs/ml/src",
  "libs/trading/src"
]
include = ["rl_trade_*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
